+++ tests/integration_tests/system/test_service_on_node.py
@@ -116,3 +116,14 @@ def test_server_turns_on_service(populated_node):
 
     assert server.operating_state is NodeOperatingState.ON
     assert service.operating_state is ServiceOperatingState.RUNNING
+
+    server.start_up_duration = 0
+    server.shut_down_duration = 0
+
+    server.power_off()
+    assert server.operating_state is NodeOperatingState.OFF
+    assert service.operating_state is ServiceOperatingState.STOPPED
+
+    server.power_on()
+    assert server.operating_state is NodeOperatingState.ON
+    assert service.operating_state is ServiceOperatingState.RUNNING